create scene
